rename - components 폴더명 setup -> filter - 필터 설정하기 url /chat/filter -> /filter
