+++ src/app/page.tsx
@@ -30,7 +30,7 @@ export default function Home() {
               추천해드릴게요. <br /> 지역, 매매 형태, 주거 형태, 예산을 선택해 주세요.
             </p>
             <Link
-              href={"/chat/filter"}
+              href={"/filter"}
               className="w-full rounded-[50px] bg-blue-50 py-2 text-center text-caption1"
             >
               필터 설정하기
